Changed package to net.snowflake.HiveMetastoreConnector, relocated package
