+++ src/main/java/net/snowflake/hivemetastoreconnector/commands/AlterExternalTable.java
@@ -1,13 +1,13 @@
 /*
  * Copyright (c) 2012-2019 Snowflake Computing Inc. All right reserved.
  */
-package com.snowflake.core.commands;
+package net.snowflake.hivemetastoreconnector.commands;
 
 import com.google.common.base.Preconditions;
 import com.google.common.collect.ImmutableList;
-import com.snowflake.conf.SnowflakeConf;
-import com.snowflake.core.util.HiveToSnowflakeType;
-import com.snowflake.core.util.StringUtil;
+import net.snowflake.hivemetastoreconnector.SnowflakeConf;
+import net.snowflake.hivemetastoreconnector.util.HiveToSnowflakeType;
+import net.snowflake.hivemetastoreconnector.util.StringUtil;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.hive.metastore.api.FieldSchema;
 import org.apache.hadoop.hive.metastore.api.Table;
