+++ src/test/java/DropPartitionTest.java
@@ -1,7 +1,7 @@
 /*
  * Copyright (c) 2012-2019 Snowflake Computing Inc. All right reserved.
  */
-import com.snowflake.core.commands.DropPartition;
+import net.snowflake.hivemetastoreconnector.commands.DropPartition;
 
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.hive.metastore.HiveMetaStore;
