+++ src/test/java/DropTableTest.java
@@ -1,7 +1,7 @@
 /*
  * Copyright (c) 2012-2019 Snowflake Computing Inc. All right reserved.
  */
-import com.snowflake.core.commands.DropExternalTable;
+import net.snowflake.hivemetastoreconnector.commands.DropExternalTable;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.hive.metastore.HiveMetaStore;
 import org.apache.hadoop.hive.metastore.api.Table;
